Turn ExternalLink into a generic Link
Closes #239
