+++ CHANGELOG.md
@@ -81,15 +81,22 @@
 
 ## `@svizzle/ui` v0.3.0 (next)
 
-- add `src/LoadingView.svelte`
-- add `src/MessageView.svelte`
-- `/src/icons/Icon.svelte`: add the `glyphSize` prop so that we can pass glyphs with sizes different from 24x24
-- add `src/icons/svizzle/A11yPerson.svelte`
-- `ScreenGauge`:
-	- add `landscape`|`portrait` to `classes`
-	- changed the output shape:
-		- remame `orientationFlags` to `orientations`
-		- remame `sizeFlags` to `sizes`
+- additions:
+	- `src/LoadingView.svelte`
+	- `src/MessageView.svelte`
+	- `src/icons/svizzle/A11yPerson.svelte`
+- changes:
+	- `Icon.svelte`: add the `glyphSize` prop so that we can pass glyphs with sizes different from 24x24
+	- turned `ExternalLink.svelte` into a more generic `Link.svelte`, adding props:
+		- `hreflang`
+		- `isDownload`
+		- `isExternal`
+		- `type`
+	- `ScreenGauge`:
+		- add `landscape`|`portrait` to `classes`
+		- changed the output shape:
+			- remame `orientationFlags` to `orientations`
+			- remame `sizeFlags` to `sizes`
 - updated some dev dependencies
 
 ## `@svizzle/utils` v0.14.0 (next)
