+++ packages/components/ui/rollup.config.js
@@ -28,9 +28,9 @@ const input = {
 	...rollupIconsInputFeather,
 	A11yPerson: 'src/icons/svizzle/A11yPerson.svelte',
 	defaults: 'src/defaults.js',
-	ExternalLink: 'src/ExternalLink.svelte',
 	Icon: 'src/icons/Icon.svelte',
 	index: 'src/index.js',
+	Link: 'src/Link.svelte',
 	LinkButton: 'src/LinkButton.svelte',
 	LoadingView: 'src/LoadingView.svelte',
 	MessageView: 'src/MessageView.svelte',
