+++ packages/components/ui/src/index.js
@@ -1,7 +1,7 @@
 export {breakpoints} from './defaults';
 export * from './gauges';
 export * from './icons';
-export {default as ExternalLink} from './ExternalLink.svelte';
+export {default as Link} from './Link.svelte';
 export {default as LinkButton} from './LinkButton.svelte';
 export {default as LoadingView} from './LoadingView.svelte';
 export {default as MessageView} from './MessageView.svelte';
